requirements.txt
python-telegram-bot==20.3
requests==2.31.0
pdfplumber==0.9.0
pytesseract==0.3.10
psycopg2-binary==2.9.6 